+++ Employees/App.xaml
@@ -1,9 +1,9 @@
-﻿<Application x:Class="lesson5.App"
+﻿<Application x:Class="Employees.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:lesson5"
+             xmlns:local="clr-namespace:Employees"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
-         
+        
     </Application.Resources>
 </Application>
