+++ Employees/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace lesson5
+namespace Employees
 {
     /// <summary>
     /// Логика взаимодействия для App.xaml
