+++ Employees/Department.xaml
@@ -1,0 +1,46 @@
+﻿<Page x:Class="Employees.Department"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:Employees"
+      mc:Ignorable="d" 
+      d:DesignHeight="600" d:DesignWidth="800"
+      Title="Подразделения">
+
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="70*"/>
+            <ColumnDefinition Width="30*"/>
+        </Grid.ColumnDefinitions>
+
+        <ListView x:Name="DeptView" Grid.Column="0" HorizontalAlignment="Left" Width="560" VirtualizingStackPanel.IsVirtualizing="True">
+            <ListView.ItemContainerStyle>
+                <Style TargetType="ListViewItem">
+                    <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
+                </Style>
+            </ListView.ItemContainerStyle>
+            <ListView.View>
+                <GridView>
+                    <GridViewColumn Header="Подразделения" Width="540">
+                        <GridViewColumn.CellTemplate>
+                            <DataTemplate>
+                                <TextBlock Text="{Binding Dept}" Foreground="Black" FontWeight="Bold" TextAlignment="Center" />
+                            </DataTemplate>
+                        </GridViewColumn.CellTemplate>
+                    </GridViewColumn>
+                    <!--<GridViewColumn Header="Подразделения" Width="228" DisplayMemberBinding="{Binding Dept}"></GridViewColumn>-->
+                </GridView>
+            </ListView.View>
+        </ListView>
+        <StackPanel Grid.Column="2">
+
+            <Label x:Name="DeptLabel" Content="Подразделение:" HorizontalContentAlignment="Center"  Margin="5" />
+            <ComboBox x:Name="DepartCombo" Margin="40,5,40,5" IsEditable="True" DisplayMemberPath="Dept" VirtualizingStackPanel.IsVirtualizing="True"/>
+            <Button x:Name="DeptbtnAdd" Margin="40,5,40,5" Content="Добавить" Click="DeptbtnAdd_Click" ToolTip="Добавление введенного подразделения"/>
+            <Button x:Name="DeptbtnDel" Margin="40,5,40,5" Content="Удалить" ToolTip="Удаление выбранного в списке подразделения" Click="DeptbtnDel_Click"/>
+            <Button x:Name="DeptbtnChg" Margin="40,5,40,5" Content="Изменить" Click="DeptbtnChg_Click" ToolTip="Изменение выбранной строки"/>
+
+        </StackPanel>
+    </Grid>
+</Page>
