+++ Employees/Employee.xaml
@@ -1,0 +1,55 @@
+﻿<Page x:Class="Employees.Employee"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:Employees"
+      mc:Ignorable="d" 
+      d:DesignHeight="600" d:DesignWidth="800"
+      Title="Сотрудники">
+
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="70*"/>
+            <ColumnDefinition Width="30*"/>
+        </Grid.ColumnDefinitions>
+
+        <ListView x:Name="EmpView" Grid.Column="0" HorizontalAlignment="Right" Width="560" VirtualizingStackPanel.IsVirtualizing="True">
+            <ListView.ItemContainerStyle>
+                <Style TargetType="ListViewItem">
+                    <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
+                </Style>
+            </ListView.ItemContainerStyle>
+            <ListView.View>
+                <GridView>
+                    <GridViewColumn Header="Сотрудник" Width="300">
+                        <GridViewColumn.CellTemplate>
+                            <DataTemplate>
+                                <TextBlock Text="{Binding EmpName}" Foreground="Black" FontWeight="Bold" TextAlignment="Center" />
+                            </DataTemplate>
+                        </GridViewColumn.CellTemplate>
+                    </GridViewColumn>
+                    <GridViewColumn Header="Подразделение" Width="245">
+                        <GridViewColumn.CellTemplate>
+                            <DataTemplate>
+                                <TextBlock Text="{Binding Dept}" Foreground="Black" FontWeight="Bold" TextAlignment="Center" />
+                            </DataTemplate>
+                        </GridViewColumn.CellTemplate>
+                    </GridViewColumn>
+                </GridView>
+            </ListView.View>
+        </ListView>
+
+        <StackPanel Grid.Column="2">
+            <Label x:Name="EmpLabel" Content="Сотрудник:" HorizontalContentAlignment="Center"  Margin="1" />            
+            <ComboBox x:Name="EmpCombo" Margin="40,5,40,5" IsEditable="True" DisplayMemberPath="EmpName" VirtualizingStackPanel.IsVirtualizing="True"/>
+            <Label x:Name="DeptLabel" Content="Подразделение:" HorizontalContentAlignment="Center"  Margin="1" />
+            <ComboBox x:Name="DeptCombo" Margin="40,5,40,5" IsEditable="True" DisplayMemberPath="Dept" VirtualizingStackPanel.IsVirtualizing="True"/>
+            <Button x:Name="EmpbtnAdd" Margin="40,5,40,5" Content="Добавить" Click="EmpbtnAdd_Click" ToolTip="Чтобы добавить надо ввести вручную сотрудника в первое поле и выбрать подразделение во 2м поле."/>
+            <Button x:Name="EmpbtnDel" Margin="40,5,40,5" Content="Удалить" Click="EmpbtnDel_Click" ToolTip="Удаление выбранного в списке сотрудника"/>
+            <Button x:Name="EmpbtnChg" Margin="40,5,40,5" Content="Изменить" Click="EmpbtnChg_Click" ToolTip="Изменение выбранной строки по введеным данным"/>
+            <Button x:Name="DeptViewBut" Margin="40,50,40,5" Content="Просмотр подразделений" Click="DeptViewBut_Click" ToolTip="Просмотр и редактирование подразделений" Height="25" />
+        </StackPanel>
+
+    </Grid>
+</Page>
