+++ Employees/Employees.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28803.452
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "lesson6", "lesson6.csproj", "{520695D2-17D7-46A6-B02A-4DD6E3495882}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Employees", "Employees.csproj", "{520695D2-17D7-46A6-B02A-4DD6E3495882}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
