+++ Employees/MainWindow.xaml
@@ -1,0 +1,6 @@
+﻿<NavigationWindow x:Class="Employees.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="Cотрудники" Height="600" Width="800" Source="Employee.xaml">
+
+</NavigationWindow>
