--- lesson5/MainWindow.xaml
@@ -1,79 +1,0 @@
-﻿<Window x:Class="lesson6.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:lesson5"
-        mc:Ignorable="d"
-        Title="Cотрудники" Height="600" Width="800">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="40*"/>
-            <ColumnDefinition Width="30*"/>
-            <ColumnDefinition Width="30*"/>
-        </Grid.ColumnDefinitions>
-
-        <ListView x:Name="EmpView" Grid.Column="0" HorizontalAlignment="Right" Width="317" VirtualizingStackPanel.IsVirtualizing="True">
-            <ListView.ItemContainerStyle>
-                <Style TargetType="ListViewItem">
-                    <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
-                </Style>
-            </ListView.ItemContainerStyle>
-            <ListView.View>
-                <GridView>
-                    <GridViewColumn Header="Сотрудники" Width="160">
-                        <GridViewColumn.CellTemplate>
-                            <DataTemplate>
-                                <TextBlock Text="{Binding Name}" Foreground="Black" FontWeight="Bold" TextAlignment="Center" />
-                            </DataTemplate>
-                        </GridViewColumn.CellTemplate>
-                    </GridViewColumn>
-                    <GridViewColumn Header="Подразд." Width="147">
-                        <GridViewColumn.CellTemplate>
-                            <DataTemplate>
-                                <TextBlock Text="{Binding Dept}" Foreground="Black" FontWeight="Bold" TextAlignment="Center" />
-                            </DataTemplate>
-                        </GridViewColumn.CellTemplate>
-                    </GridViewColumn>
-                </GridView>
-            </ListView.View>
-        </ListView>
-
-        <ListView x:Name="DeptView" Grid.Column="1" HorizontalAlignment="Left" Width="238" VirtualizingStackPanel.IsVirtualizing="True">
-            <ListView.ItemContainerStyle>
-                <Style TargetType="ListViewItem">
-                    <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
-                </Style>
-            </ListView.ItemContainerStyle>
-            <ListView.View>
-                <GridView>
-                    <GridViewColumn Header="Подразделения" Width="228">
-                        <GridViewColumn.CellTemplate>
-                            <DataTemplate>
-                                <TextBlock Text="{Binding Dept}" Foreground="Black" FontWeight="Bold" TextAlignment="Center" />
-                            </DataTemplate>
-                        </GridViewColumn.CellTemplate>
-                    </GridViewColumn>
-                    <!--<GridViewColumn Header="Подразделения" Width="228" DisplayMemberBinding="{Binding Dept}"></GridViewColumn>-->
-                </GridView>
-            </ListView.View>
-        </ListView>
-
-        <StackPanel Grid.Column="2">
-            <Label x:Name="EmpLabel" Content="Сотрудник:" HorizontalContentAlignment="Center"  Margin="5" />
-            <ComboBox x:Name="EmpCombo" Margin="40,5,40,5" IsEditable="True" DisplayMemberPath="Name" VirtualizingStackPanel.IsVirtualizing="True"/>
-            <ComboBox x:Name="DeptCombo" Margin="40,5,40,5" IsEditable="True" DisplayMemberPath="Dept" VirtualizingStackPanel.IsVirtualizing="True"/>                        
-            <Button x:Name="EmpbtnAdd" Margin="40,5,40,5" Content="Добавить" Click="EmpbtnAdd_Click" ToolTip="Чтобы добавить надо ввести вручную сотрудника в первое поле и выбрать подразделение во 2м поле."/>
-            <Button x:Name="EmpbtnDel" Margin="40,5,40,5" Content="Удалить" Click="EmpbtnDel_Click" ToolTip="Удаление выбранного в списке сотрудника"/>
-            <Button x:Name="EmpbtnChg" Margin="40,5,40,5" Content="Изменить" Click="EmpbtnChg_Click" ToolTip="Изменение выбранной строки"/>
-
-            <Label x:Name="DeptLabel" Content="Подразделение:" HorizontalContentAlignment="Center"  Margin="5" />
-            <ComboBox x:Name="DepartCombo" Margin="40,5,40,5" IsEditable="True" DisplayMemberPath="Dept" VirtualizingStackPanel.IsVirtualizing="True"/>
-            <Button x:Name="DeptbtnAdd" Margin="40,5,40,5" Content="Добавить" Click="DeptbtnAdd_Click" ToolTip="Добавление введенного подразделения"/>
-            <Button x:Name="DeptbtnDel" Margin="40,5,40,5" Content="Удалить" ToolTip="Удаление выбранного в списке подразделения" Click="DeptbtnDel_Click"/>
-            <Button x:Name="DeptbtnChg" Margin="40,5,40,5" Content="Изменить" Click="DeptbtnChg_Click" ToolTip="Изменение выбранной строки"/>
-
-        </StackPanel>
-
-    </Grid>
-</Window>
